web_cms/requirements.txt
Mezzanine==3.1.4
mysql-python
django-auth-ldap==1.2.0
django-markup-deprecated==0.0.3
